+++ app/src/main/kotlin/com/simplemobiletools/contacts/activities/SelectContactActivity.kt
@@ -6,9 +6,7 @@ import android.os.Bundle
 import android.provider.ContactsContract
 import android.view.Menu
 import android.view.MenuItem
-import com.simplemobiletools.commons.extensions.baseConfig
-import com.simplemobiletools.commons.extensions.isActivityDestroyed
-import com.simplemobiletools.commons.extensions.toast
+import com.simplemobiletools.commons.extensions.*
 import com.simplemobiletools.commons.helpers.PERMISSION_READ_CONTACTS
 import com.simplemobiletools.commons.helpers.PERMISSION_WRITE_CONTACTS
 import com.simplemobiletools.contacts.R
@@ -21,14 +19,15 @@ import com.simplemobiletools.contacts.extensions.getVisibleContactSources
 import com.simplemobiletools.contacts.helpers.ContactsHelper
 import com.simplemobiletools.contacts.helpers.SMT_PRIVATE
 import com.simplemobiletools.contacts.models.Contact
-import kotlinx.android.synthetic.main.layout_select_contact.*
+import kotlinx.android.synthetic.main.activity_select_contact.*
 
 class SelectContactActivity : SimpleActivity() {
     private var specialMimeType: String? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.layout_select_contact)
+        setContentView(R.layout.activity_select_contact)
+        setupPlaceholders()
 
         handlePermission(PERMISSION_READ_CONTACTS) {
             if (it) {
@@ -80,7 +79,7 @@ class SelectContactActivity : SimpleActivity() {
 
     private fun initContacts() {
         ContactsHelper(this).getContacts {
-            if (isActivityDestroyed()) {
+            if (isDestroyed) {
                 return@getContacts
             }
 
@@ -105,8 +104,11 @@ class SelectContactActivity : SimpleActivity() {
             contacts.sort()
 
             runOnUiThread {
-                select_contact_list.adapter = SelectContactsAdapter(this, contacts, ArrayList(), false) {
+                updatePlaceholderVisibility(contacts)
+                SelectContactsAdapter(this, contacts, ArrayList(), false, select_contact_list) {
                     confirmSelection(it)
+                }.apply {
+                    select_contact_list.adapter = this
                 }
 
                 select_contact_fastscroller.allowBubbleDisplay = baseConfig.showInfoBubble
@@ -132,9 +134,29 @@ class SelectContactActivity : SimpleActivity() {
                 val contactId = ContactsHelper(this).getContactMimeTypeId(contact.id.toString(), specialMimeType!!)
                 Uri.withAppendedPath(ContactsContract.Data.CONTENT_URI, contactId)
             }
-            else -> {
-                getContactPublicUri(contact)
+            else -> getContactPublicUri(contact)
+        }
+    }
+
+    private fun setupPlaceholders() {
+        select_contact_placeholder.setTextColor(config.textColor)
+        select_contact_placeholder_2.setTextColor(getAdjustedPrimaryColor())
+        select_contact_placeholder_2.underlineText()
+        select_contact_placeholder_2.setOnClickListener {
+            FilterContactSourcesDialog(this) {
+                initContacts()
             }
         }
+    }
+
+    private fun updatePlaceholderVisibility(contacts: ArrayList<Contact>) {
+        select_contact_list.beVisibleIf(contacts.isNotEmpty())
+        select_contact_placeholder_2.beVisibleIf(contacts.isEmpty())
+        select_contact_placeholder.beVisibleIf(contacts.isEmpty())
+        select_contact_placeholder.setText(when (specialMimeType) {
+            ContactsContract.CommonDataKinds.Email.CONTENT_ITEM_TYPE -> R.string.no_contacts_with_emails
+            ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE -> R.string.no_contacts_with_phone_numbers
+            else -> R.string.no_contacts_found
+        })
     }
 }