+++ app/src/main/kotlin/com/simplemobiletools/contacts/adapters/SelectContactsAdapter.kt
@@ -1,10 +1,11 @@
 package com.simplemobiletools.contacts.adapters
 
 import android.graphics.drawable.Drawable
-import android.support.v7.widget.RecyclerView
 import android.util.SparseArray
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
@@ -13,8 +14,7 @@ import com.bumptech.glide.signature.ObjectKey
 import com.simplemobiletools.commons.extensions.beVisibleIf
 import com.simplemobiletools.commons.extensions.getAdjustedPrimaryColor
 import com.simplemobiletools.commons.extensions.getColoredDrawableWithColor
-import com.simplemobiletools.commons.extensions.isActivityDestroyed
-import com.simplemobiletools.commons.interfaces.MyAdapterListener
+import com.simplemobiletools.commons.views.MyRecyclerView
 import com.simplemobiletools.contacts.R
 import com.simplemobiletools.contacts.activities.SimpleActivity
 import com.simplemobiletools.contacts.extensions.config
@@ -24,7 +24,7 @@ import kotlinx.android.synthetic.main.item_add_favorite_with_number.view.*
 import java.util.*
 
 class SelectContactsAdapter(val activity: SimpleActivity, val contacts: List<Contact>, private val selectedContacts: ArrayList<Contact>, private val allowPickMultiple: Boolean,
-                            private val itemClick: ((Contact) -> Unit)? = null) : RecyclerView.Adapter<SelectContactsAdapter.ViewHolder>() {
+                            private val recyclerView: MyRecyclerView, private val itemClick: ((Contact) -> Unit)? = null) : RecyclerView.Adapter<SelectContactsAdapter.ViewHolder>() {
     private val itemViews = SparseArray<View>()
     private val selectedPositions = HashSet<Int>()
     private val config = activity.config
@@ -38,9 +38,18 @@ class SelectContactsAdapter(val activity: SimpleActivity, val contacts: List<Con
 
     init {
         contacts.forEachIndexed { index, contact ->
-            if (selectedContacts.map { it.id }.contains(contact.id)) {
+            if (selectedContacts.asSequence().map { it.id }.contains(contact.id)) {
                 selectedPositions.add(index)
             }
+        }
+
+        if (recyclerView.itemDecorationCount > 0) {
+            recyclerView.removeItemDecorationAt(0)
+        }
+
+        DividerItemDecoration(activity, DividerItemDecoration.VERTICAL).apply {
+            setDrawable(activity.resources.getDrawable(R.drawable.divider))
+            recyclerView.addItemDecoration(this)
         }
     }
 
@@ -56,16 +65,6 @@ class SelectContactsAdapter(val activity: SimpleActivity, val contacts: List<Con
         itemViews[pos]?.contact_checkbox?.isChecked = select
     }
 
-    private val adapterListener = object : MyAdapterListener {
-        override fun toggleItemSelectionAdapter(select: Boolean, position: Int) {
-            toggleItemSelection(select, position)
-        }
-
-        override fun getSelectedPositions() = selectedPositions
-
-        override fun itemLongClicked(position: Int) {}
-    }
-
     fun getSelectedItemsSet(): HashSet<Contact> {
         val selectedItemsSet = HashSet<Contact>(selectedPositions.size)
         selectedPositions.forEach { selectedItemsSet.add(contacts[it]) }
@@ -74,7 +73,7 @@ class SelectContactsAdapter(val activity: SimpleActivity, val contacts: List<Con
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = activity.layoutInflater.inflate(itemLayout, parent, false)
-        return ViewHolder(view, adapterListener, activity, allowPickMultiple, itemClick)
+        return ViewHolder(view)
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
@@ -85,16 +84,15 @@ class SelectContactsAdapter(val activity: SimpleActivity, val contacts: List<Con
 
     override fun getItemCount() = contacts.size
 
-    class ViewHolder(view: View, private val adapterListener: MyAdapterListener, val activity: SimpleActivity, private val showCheckbox: Boolean,
-                     private val itemClick: ((Contact) -> Unit)?) : RecyclerView.ViewHolder(view) {
+    inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         fun bindView(contact: Contact, contactDrawable: Drawable, config: Config, showContactThumbnails: Boolean,
                      smallPadding: Int, bigPadding: Int): View {
             itemView.apply {
-                contact_checkbox.beVisibleIf(showCheckbox)
+                contact_checkbox.beVisibleIf(allowPickMultiple)
                 contact_checkbox.setColors(config.textColor, context.getAdjustedPrimaryColor(), config.backgroundColor)
                 val textColor = config.textColor
 
-                contact_name.text = contact.getFullName()
+                contact_name.text = contact.getNameToDisplay()
                 contact_name.setTextColor(textColor)
                 contact_name.setPadding(if (showContactThumbnails) smallPadding else bigPadding, smallPadding, smallPadding, 0)
 
@@ -119,7 +117,7 @@ class SelectContactsAdapter(val activity: SimpleActivity, val contacts: List<Con
                                 .error(contactDrawable)
                                 .centerCrop()
 
-                        if (activity.isActivityDestroyed()) {
+                        if (activity.isDestroyed) {
                             Glide.with(activity).load(contact.photoUri).transition(DrawableTransitionOptions.withCrossFade()).apply(options).into(contact_tmb)
                         }
                     } else {
@@ -132,14 +130,14 @@ class SelectContactsAdapter(val activity: SimpleActivity, val contacts: List<Con
         }
 
         private fun viewClicked(select: Boolean) {
-            adapterListener.toggleItemSelectionAdapter(select, adapterPosition)
+            toggleItemSelection(select, adapterPosition)
         }
     }
 
     override fun onViewRecycled(holder: ViewHolder) {
         super.onViewRecycled(holder)
-        if (!activity.isActivityDestroyed()) {
-            Glide.with(activity).clear(holder.itemView?.contact_tmb!!)
+        if (!activity.isDestroyed) {
+            Glide.with(activity).clear(holder.itemView.contact_tmb)
         }
     }
 }