+++ app/src/main/kotlin/com/simplemobiletools/contacts/dialogs/RenameGroupDialog.kt
@@ -1,6 +1,6 @@
 package com.simplemobiletools.contacts.dialogs
 
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 import com.simplemobiletools.commons.activities.BaseSimpleActivity
 import com.simplemobiletools.commons.extensions.*
 import com.simplemobiletools.contacts.R