+++ app/src/main/kotlin/com/simplemobiletools/contacts/extensions/Activity.kt
@@ -36,7 +36,7 @@ fun SimpleActivity.startCallIntent(recipient: String) {
 
 fun SimpleActivity.tryStartCall(contact: Contact) {
     if (config.showCallConfirmation) {
-        CallConfirmationDialog(this, contact.getFullName()) {
+        CallConfirmationDialog(this, contact.getNameToDisplay()) {
             startCall(contact)
         }
     } else {
@@ -63,8 +63,9 @@ fun SimpleActivity.startCall(contact: Contact) {
 fun SimpleActivity.showContactSourcePicker(currentSource: String, callback: (newSource: String) -> Unit) {
     ContactsHelper(this).getContactSources {
         val ignoredTypes = arrayListOf(
-                "org.thoughtcrime.securesms",   // Signal
-                "org.telegram.messenger"        // Telegram
+                SIGNAL_PACKAGE,
+                TELEGRAM_PACKAGE,
+                WHATSAPP_PACKAGE
         )
 
         val items = ArrayList<RadioItem>()