+++ app/src/main/kotlin/com/simplemobiletools/contacts/helpers/VcfImporter.kt
@@ -139,7 +139,7 @@ class VcfImporter(val activity: SimpleActivity) {
                         targetContactSource, starred, contactId, thumbnailUri, photo, notes, groups, organization, websites, cleanPhoneNumbers, IMs)
 
                 // if there is no N and ORG fields at the given contact, only FN, treat it as an organization
-                if (contact.getFullName().isEmpty() && contact.organization.isEmpty() && ezContact.formattedName.value.isNotEmpty()) {
+                if (contact.getNameToDisplay().isEmpty() && contact.organization.isEmpty() && ezContact.formattedName.value.isNotEmpty()) {
                     contact.organization.company = ezContact.formattedName.value
                 }